Major refactoring of code
