+++ .gitignore
@@ -2,3 +2,4 @@
 node_modules/
 backend/*/dist/
 backend/*/coverage/
+**/.DS_Store
