+++ frontend/package.json
@@ -20,6 +20,7 @@
     "@radix-ui/react-select": "^2.1.1",
     "@radix-ui/react-separator": "^1.1.0",
     "@radix-ui/react-slot": "^1.1.0",
+    "@radix-ui/react-tooltip": "^1.1.3",
     "@react-oauth/google": "^0.12.1",
     "@stomp/stompjs": "^7.0.0",
     "@tanstack/react-table": "^8.20.5",
