+++ frontend/src/api/user.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable */
 
 import toast from "@/components/modals/toast";
-import { UserLogin } from "@/types/user";
+import { AuthStatus } from "@/types/user";
 import Cookie from "js-cookie";
 
 export const setToken = (token: string) => {
@@ -12,33 +12,39 @@ export const getToken = () => {
   return Cookie.get("token");
 };
 
-export const setBaseUserData = (data: {
-  username: string;
-  id: string;
-  isAdmin: string;
-}) => {
-  Cookie.set("username", data.username, { expires: 1 });
-  Cookie.set("id", data.id, { expires: 1 });
-  Cookie.set("isAdmin", data.isAdmin, { expires: 1 });
-};
+export const setUsername = (username: string) => {
+  Cookie.set("username", username, { expires: 1 });
+}
 
-export const getBaseUserData = () => {
-  return {
-    username: Cookie.get("username"),
-    id: Cookie.get("id"),
-    isAdmin: Cookie.get("isAdmin"),
-  };
-};
+export const getUsername = () => {
+  return Cookie.get("username");
+}
+
+export const setUserId = (id: string) => {
+  Cookie.set("id", id, { expires: 1 });
+}
+
+export const getUserId = () => {
+  return Cookie.get("id");
+}
+
+export const setIsAdmin = (isAdmin: string) => {
+  Cookie.set("isAdmin", isAdmin, { expires: 1 });
+}
+
+export const getIsAdmin = () => {
+  return Cookie.get("isAdmin");
+}
+
+export const getAuthStatus = () => {
+  if (!getToken()) return AuthStatus.UNAUTHENTICATED;
+  if (getIsAdmin() === "true") return AuthStatus.ADMIN;
+  return AuthStatus.AUTHENTICATED;
+}
 
 const NEXT_PUBLIC_USER_SERVICE = process.env.NEXT_PUBLIC_USER_SERVICE || "https://user-service-598285527681.us-central1.run.app";
 
-export const verifyToken = async (needsLogin: boolean) => {
-  const token = getToken();
-  if (!token) {
-    if (needsLogin) logout();
-    return false;
-  }
-
+export const verifyToken = async (token: string) => {
   const response = await fetch(
     `${NEXT_PUBLIC_USER_SERVICE}/auth/verify-token`,
     {
@@ -49,14 +55,7 @@ export const verifyToken = async (needsLogin: boolean) => {
     }
   );
 
-  switch (response.status) {
-    case 200:
-      return true;
-    case 401:
-      return false;
-    default:
-      return false;
-  }
+  return response.status === 200;
 };
 
 export const login = async (email: string, password: string) => {
@@ -72,36 +71,20 @@ export const login = async (email: string, password: string) => {
   });
   const data = await response.json();
 
-  switch (response.status) {
-    case 200:
-      handleSuccessfulLogin(data.data);
-      break;
-    default:
-      toast.fire({
-        icon: "error",
-        title: data.message,
-      });
-  }
-};
-
-export const handleSuccessfulLogin = async (data: UserLogin) => {
-  const { accessToken, username, id, isAdmin } = data;
-  setBaseUserData({ username, id, isAdmin });
-  setToken(accessToken);
-  const redirect = Cookie.get("redirect");
-  if (!redirect) {
-    window.location.href = "/";
-    return;
+  if (response.status !== 200) {
+    toast.fire({
+      icon: "error",
+      title: data.message,
+    });
+    return false;
   }
 
-  Cookie.remove("redirect");
-  window.location.href = redirect;
-  return;
-};
+  setToken(data.data.accessToken);
+  setUsername(data.data.username);
+  setIsAdmin(data.data.isAdmin);
+  setUserId(data.data.id);
 
-export const redirectToLogin = async () => {
-  Cookie.set("redirect", window.location.pathname, { expires: 1 });
-  window.location.href = "/login";
+  return true;
 };
 
 export const logout = () => {
@@ -109,7 +92,8 @@ export const logout = () => {
   Cookie.remove("username");
   Cookie.remove("id");
   Cookie.remove("isAdmin");
-  redirectToLogin();
+
+  window.location.href = "/";
 };
 
 export const register = async (
@@ -130,43 +114,42 @@ export const register = async (
   });
   const data = await response.json();
 
-  switch (response.status) {
-    case 201:
-      return handleSuccessfulLogin(data.data);
-    default:
-      toast.fire({
-        icon: "error",
-        title: data.message,
-      });
+  if (response.status !== 201) {
+    toast.fire({
+      icon: "error",
+      title: data.message,
+    });
+    return false;
   }
+
+  return true;
 };
 
-export const getUser = async () => {
+// optional userId parameter
+export const getUser = async (userId = "") => {
   const token = getToken();
-  const { id } = getBaseUserData();
-  const response = await fetch(`${NEXT_PUBLIC_USER_SERVICE}/users/${id}`, {
+  const url = `${NEXT_PUBLIC_USER_SERVICE}/users/${userId || getUserId()}`
+  console.log(url);
+  const response = await fetch(url, {
     method: "GET",
     headers: {
       Authorization: `Bearer ${token}`,
     },
   });
+  const data = await response.json();
 
-  switch (response.status) {
-    case 200:
-      return await response.json();
-    case 401:
-      logout();
-      break;
-    default:
-      const { message } = await response.json();
-      toast.fire({
-        icon: "error",
-        title: message,
-      });
+  if (response.status !== 200) {
+    toast.fire({
+      icon: "error",
+      title: data.message,
+    });
+    return false;
   }
+
+  return data;
 };
 
-export const updateUser = async (data: {
+export const updateUser = async (userData: {
   username?: string;
   email?: string;
   password?: string;
@@ -175,29 +158,29 @@ export const updateUser = async (data: {
   github?: string;
 }) => {
   const token = getToken();
-  const { id } = getBaseUserData();
-  const response = await fetch(`${NEXT_PUBLIC_USER_SERVICE}/users/${id}`, {
+  const userId = getUserId();
+  const response = await fetch(`${NEXT_PUBLIC_USER_SERVICE}/users/${userId}`, {
     method: "PATCH",
     headers: {
       "Content-Type": "application/json",
       Authorization: `Bearer ${token}`,
     },
-    body: JSON.stringify(data),
+    body: JSON.stringify(userData),
+  });
+  const data = await response.json();
+
+  if (response.status !== 200) {
+    toast.fire({
+      icon: "error",
+      title: data.message,
+    });
+    return false;
+  }
+
+  toast.fire({
+    icon: "success",
+    title: "Profile updated successfully",
   });
 
-  switch (response.status) {
-    case 200:
-      toast.fire({
-        icon: "success",
-        title: "Profile updated successfully",
-      });
-      break;
-    default:
-      // get the error message
-      const { message } = await response.json();
-      toast.fire({
-        icon: "error",
-        title: message,
-      });
-  }
+  return true;
 };
