+++ frontend/src/app/(user)/dashboard/page.tsx
@@ -1,0 +1,7 @@
+import Dashboard from "@/app/(user)/dashboard/components/Dashboard"
+
+const DashboardPage = () => {
+  return <Dashboard />;
+}
+
+export default DashboardPage;
