+++ frontend/src/app/(user)/login/page.tsx
@@ -1,0 +1,7 @@
+import Login from "@/app/(user)/login/components/Login";
+
+const LoginPage = () => {
+  return <Login/>;
+};
+
+export default LoginPage;
