+++ frontend/src/app/(user)/register/page.tsx
@@ -1,0 +1,7 @@
+import Register from "@/app/(user)/register/components/Register";
+
+const RegisterPage = () => {
+  return <Register/>;
+};
+
+export default RegisterPage;
