+++ frontend/src/app/admin/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import QuestionDashboard from "@/app/admin/components/QuestionDashboard";
+
+const QuestionDashboardPage = () => {
+  return <QuestionDashboard />;
+};
+
+export default QuestionDashboardPage;
