+++ frontend/src/app/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import LandingPage from "@/app/(home)/components/landing-page/LandingPage";
+
+const Home = () => {
+  return <LandingPage />;
+};
+
+export default Home;
