+++ frontend/src/types/dashboard.tsx
@@ -1,3 +1,17 @@
+export type TSession = {
+  collabid: string;
+  users: [string];
+  language: [string];
+  question_id: number;
+};
+
+export type TQuestion = {
+  questionid: number;
+  title: string;
+  complexity: string;
+  category: string;
+};
+
 export type TCombinedSession = {
   collabid: string;
   peer: string;
