+++ frontend/src/types/user.ts
@@ -1,3 +1,10 @@
+export enum AuthStatus {
+  LOADING = "LOADING",
+  UNAUTHENTICATED = "UNAUTHENTICATED",
+  AUTHENTICATED = "AUTHENTICATED",
+  ADMIN = "ADMIN",
+}
+
 export interface AttemptedQuestion {
   title: string;
   peer: string;
@@ -11,6 +18,7 @@ export interface User {
   bio?: string;
   linkedin?: string;
   github?: string;
+  createdAt?: string;
 }
 
 export interface UserLogin {
